Move new feature out from behind dev flag
