+++ hpc_provisioner/src/hpc_provisioner/config-dev/_head_node.tpl.yaml
@@ -17,7 +17,7 @@ Iam:
 CustomActions:
   OnNodeConfigured:
     Sequence:
-      - Script: s3://sboinfrastructureassets-sandbox/scripts/dev_create_users.py
+      - Script: s3://sboinfrastructureassets-sandbox/scripts/create_users.py
         Args: !config create_users_args
 #       - Script: s3://sboinfrastructureassets/scripts/all_or_nothing_allocation.sh
 #       - Script: s3://sboinfrastructureassets/scripts/setup_lustre_directories.py
